pull out all the isolated subcomponents
